+++ docs/changelog.rst
@@ -27,6 +27,7 @@ Changes in 0.9.X - DEV
 - Removed get_or_create() deprecated since 0.8.0. #300
 - Capped collection multiple of 256. #1011
 - Added `BaseQuerySet.aggregate_sum` and `BaseQuerySet.aggregate_average` methods.
+- Fix for delete with write_concern {'w': 0}. #1008
 
 Changes in 0.9.0
 ================